Beam Database query tool change to save all entries (#274)
If database entry for a particular device doesn't return any values, fill that device branch in the output root file with -9999 instead of leaving it empty. 

We use this tool in the BeamFetcherV2 toolchain to produce beam root files that are then used by the event building toolchain. The beam root files contain a branch for each device. Without this modification, as the BeamFetcherV2 toolchain loops over RAWData part files, grabs CTC timestamps and queries for DB values, if a device entry is empty the toolchain will fail to fill the entire device branch as the entries between branches will be unequal.